extract Alien class to separate file
