+++ models.py
@@ -1,39 +1,8 @@
-import random
 from typing import List
 
 import pygame as pg
 
 from settings import Settings
-
-
-class Alien(pg.sprite.Sprite):
-    """An alien space ship. That slowly moves down the screen."""
-
-    speed = 13
-    animcycle = 12
-    images: List[pg.Surface] = []
-
-    def __init__(self, *groups):
-
-        self.settings = Settings()
-        self.SCREENRECT = self.settings.SCREENRECT
-
-        pg.sprite.Sprite.__init__(self, *groups)
-        self.image = self.images[0]
-        self.rect = self.image.get_rect()
-        self.facing = random.choice((-1, 1)) * Alien.speed
-        self.frame = 0
-        if self.facing < 0:
-            self.rect.right = self.SCREENRECT.right
-
-    def update(self):
-        self.rect.move_ip(self.facing, 0)
-        if not self.SCREENRECT.contains(self.rect):
-            self.facing = -self.facing
-            self.rect.top = self.rect.bottom + 1
-            self.rect = self.rect.clamp(self.SCREENRECT)
-        self.frame = self.frame + 1
-        self.image = self.images[self.frame // self.animcycle % 3]
 
 
 class Bomb(pg.sprite.Sprite):
